Fixed various grammar errors & typos
Those are very minor errors, probably from the translator's inattention. Nonetheless, some of these looked very obvious.

Line #177 - Because "Filenames" is one plural word and that the translation has to be a noun group (There is no french word for "filename", and "nomdefichier", "nomfichier" or "fichiernom" all sound awful and don't even exist), we need to conjugate all countable elements from the noun group to it's central noun which, in this case, is "Noms".
Line #178 - Lacked an accent.
Line #201 - 3 things here : 
     1 - "Choose" was conjugated without an "s" at the end, meaning it wasn't using the 3rd singular pronoun from the present verb tense, it used the 1st form of the imperative verb tense. In French, it's gonna be "Choisis" instead of "Choisit".
     2 - "Accéder", not "accède". Here, it's the opposite of the issue on line #335. I could give an explanation on that, but honestly, I don't think it's necessary, plus it's so complicated that I have trouble explaining it. I could try giving one though if it's requested. 
     3 - Correct me if I'm wrong, but I felt that changing that "Bibliothèques" part with french quotation marks and removing the "de" sounds grammatically better.
Line #217 - "Épingler" has an accent on the first "E".(See http://prntscr.com/3ltvrf if you don't believe me)
Line #220 - See the identical problem above.
Line #221 - "pivoter" needs an "i", just like the line after it.
Line #241 - See #217.
Line #299 - "^" accent was on the wrong vowel, 'nuff said.
Line #335 - When talking to a user, it's usually better to use the present indicative verb tense with the 2nd plural pronoum (Just like in English) instead of the infinitive one that was used in these examples, so I took the liberty to change those to the former-mentioned verb tense.
Lines #365, #366, #367 & #368 - Needs a "^" accent on "Etes-vous".
Line #383 - One "t" only in "Sélectionner". Was probably due to inattention.
Line #400 & #401 - Unneeded anglicism. There already exists a word for "Customize".

Hopefully I won't have wasted my time helping to translate this, and I appologize if it's hard to read. I might check the typos some more if I have time to.